binding: expose tw_binding_init for intergrating bingings into config.
